Pull splitter logic into a splitter module
